moved unit test to the appropriate location
I overlooked the presence of an existing `tests.py` file.  My bad.
